@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pytest-zebrunner"
-version = "0.1.1"
+version = "1.6.2"
 description = "Pytest connector for Zebrunner reporting"
 authors = ["Anatoliy Platonov <p4m.dev@gmail.com>"]
 license = "Apache"
@@ -18,12 +18,12 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
     "Topic :: Software Development :: Libraries",
     "Topic :: Software Development",
-    "Development Status :: 4 - Beta",
+    "Development Status :: 5 - Production/Stable",
     "License :: OSI Approved :: Apache Software License",
     "Framework :: Pytest",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
 ]
 
 
@@ -32,26 +32,21 @@ pytest-zebrunner = "pytest_zebrunner.plugin"
 
 
 [tool.poetry.dependencies]
-python = "^3.7"
-pytest = "6.0.2"
-httpx = "^0.16.1"
-pydantic = "^1.7.2"
-python-dotenv = "^0.15.0"
-PyYAML = "^5.4.1"
-
+python = ">=3.8,<4.0"
+pytest = ">=4.5.0"
+httpx = "^0.25.0"
+pydantic = "^1.0"
+python-dotenv = ">=0.10"
+PyYAML = "^6.0.1"
 
 [tool.poetry.dev-dependencies]
-black = "^20.8b1"
+black = "^23.9.1"
 isort = "^5.6.4"
-mypy = "^0.790"
-pre-commit = "^2.7.1"
-pytest-cov = "^2.10.1"
-selenium = "^3.141.0"
-setuptools = "^56.0.0"
-pytest-xdist = "^2.2.1"
-
-[tool.poetry.extras]
-xdist = ["pytest-xdist"]
+mypy = "^1.5.1"
+pre-commit = "^3.4.0"
+pytest-cov = "^4.1.0"
+selenium = "^4.0"
+setuptools = "^68.2.2"
 
 [tool.coverage.paths]
 source = ["src", "*/site-packages"]
