@@ -8,7 +8,10 @@ from pytest_zebrunner.attachments import (
     attach_test_screenshot,
 )
 
-__version__ = "0.1.1"
+from .current_test import CurrentTest  # noqa: F401
+from .current_test_run import CurrentTestRun  # noqa: F401
+
+__version__ = "1.5.6"
 __all__ = [
     "attach_test_artifact_reference",
     "attach_test_run_artifact_reference",
