@@ -5,13 +5,6 @@ from typing import Dict, List, Optional, Type
 from pytest_zebrunner.api.models import CiContextModel
 
 
-class BaseContextLoader:
-    @staticmethod
-    def load_context_variables(prefixes: List[str]) -> Dict[str, str]:
-        env_variable_names = list(filter(lambda name: any([name.startswith(x) for x in prefixes]), os.environ))
-        return {key: os.environ[key] for key in env_variable_names}
-
-
 class CiType(Enum):
     JENKINS = "JENKINS"
     TEAM_CITY = "TEAM_CITY"
@@ -20,7 +13,17 @@ class CiType(Enum):
     BAMBOO = "BAMBOO"
 
 
-class JenkinsContextLoader(BaseContextLoader):
+class CiVariables:
+    CI_ENV_VARIABLE: str
+    CI_TYPE: CiType
+    ENV_VARIABLE_PREFIXES: List[str]
+
+
+class Jenkins(CiVariables):
+    """
+    Jenkins env variables
+    """
+
     CI_ENV_VARIABLE = "JENKINS_URL"
     CI_TYPE = CiType.JENKINS
     ENV_VARIABLE_PREFIXES = [
@@ -37,15 +40,12 @@ class JenkinsContextLoader(BaseContextLoader):
         "WORKSPACE",
     ]
 
-    @classmethod
-    def resolve(cls) -> Optional[Dict[str, str]]:
-        if cls.CI_ENV_VARIABLE in os.environ:
-            return cls.load_context_variables(cls.ENV_VARIABLE_PREFIXES)
-        else:
-            return None
 
+class TeamCity(CiVariables):
+    """
+    TeamCity environemt variables
+    """
 
-class TeamCityCiContextResolver(BaseContextLoader):
     CI_ENV_VARIABLE = "TEAMCITY_VERSION"
     CI_TYPE = CiType.TEAM_CITY
     ENV_VARIABLE_PREFIXES = [
@@ -55,56 +55,57 @@ class TeamCityCiContextResolver(BaseContextLoader):
         "TEAMCITY_",
     ]
 
-    @classmethod
-    def resolve(cls) -> Optional[Dict[str, str]]:
-        if cls.CI_ENV_VARIABLE in os.environ:
-            return cls.load_context_variables(cls.ENV_VARIABLE_PREFIXES)
-        else:
-            return None
 
+class CircleCi(CiVariables):
+    """
+    CircleCI environemt variables
+    """
 
-class CircleCiContextResolver(BaseContextLoader):
     CI_ENV_VARIABLE = "CIRCLECI"
     CI_TYPE = CiType.CIRCLE_CI
     ENV_VARIABLE_PREFIXES = ["CIRCLE", "HOSTNAME"]
 
-    @classmethod
-    def resolve(cls) -> Optional[Dict[str, str]]:
-        if cls.CI_ENV_VARIABLE in os.environ:
-            return cls.load_context_variables(cls.ENV_VARIABLE_PREFIXES)
-        else:
-            return None
 
+class TravisCi(CiVariables):
+    """
+    CircleCI environemt variables
+    """
 
-class TravisCiContextResolver(BaseContextLoader):
     CI_ENV_VARIABLE = "TRAVIS"
     CI_TYPE = CiType.TRAVIS_CI
     ENV_VARIABLE_PREFIXES = ["TRAVIS", "USER"]
 
-    @classmethod
-    def resolve(cls) -> Optional[Dict[str, str]]:
-        if cls.CI_ENV_VARIABLE in os.environ:
-            return cls.load_context_variables(cls.ENV_VARIABLE_PREFIXES)
-        else:
-            return None
 
+class CiContextLoader:
+    @staticmethod
+    def load_context_variables(prefixes: List[str]) -> Dict[str, str]:
+        """
+        Returns a dictionary with environment variables that start with one of prefixes
+        """
+        env_variable_names = list(filter(lambda name: any([name.startswith(x) for x in prefixes]), os.environ))
+        return {key: os.environ[key] for key in env_variable_names}
 
-def resolve_ci_context() -> Optional[CiContextModel]:
-    ci_tools: List[Type[BaseContextLoader]] = [
-        JenkinsContextLoader,
-        TeamCityCiContextResolver,
-        CircleCiContextResolver,
-        TravisCiContextResolver,
-    ]
+    @staticmethod
+    def resolve_ci_context() -> Optional[CiContextModel]:
+        """
+        Go through list of ci variables definitions and try to load variables related to this definition.
+        Return first found ci context
+        """
 
-    ci_context: Optional[Type] = None
-    for resolver in ci_tools:
-        env_variables = resolver.resolve()  # type: ignore
-        if env_variables:
-            ci_context = resolver
-            break
+        ci_services_variables: List[Type[CiVariables]] = [
+            Jenkins,
+            TeamCity,
+            CircleCi,
+            TravisCi,
+        ]
 
-    if ci_context:
-        return CiContextModel(ci_type=ci_context.CI_TYPE.value, env_variables=ci_context.resolve())
-    else:
+        for ci_variables in ci_services_variables:
+            if ci_variables.CI_ENV_VARIABLE not in os.environ:
+                continue
+
+            return CiContextModel(
+                ci_type=ci_variables.CI_TYPE.value,
+                env_variables=CiContextLoader.load_context_variables(ci_variables.ENV_VARIABLE_PREFIXES),
+            )
+
         return None
